+++ css/styles.css
@@ -80,10 +80,11 @@ body {
   position: fixed;
   color: pink;
   width: 100%;
-  height: 90px;
+  height: 55px;
   top: 0;
   left: 0;
   right: 0;
+  z-index: 3;
 }
 
 /* -------------------------------------------------------------------------------------------------------------------------- */
@@ -97,7 +98,8 @@ body {
   align-items: center;
   justify-content: center;
   background-color: white;
-  padding: 50px 0;
+  padding-top: 10px;
+  padding-bottom: 30px;
 }
 
 .header {
@@ -154,7 +156,7 @@ body {
 
 .hamburger {
   position: absolute;
-  top: 54px;
+  top: 15px;
   right: 24px;
 }
 
@@ -162,10 +164,10 @@ body {
   font-family: "Inter", sans-serif;
   font-style: normal;
   font-weight: 600;
-  font-size: 20px;
+  font-size: 15px;
   color: #28352f;
   position: absolute;
-  top: 54px;
+  top: 15px;
   left: 24px;
 }
 
@@ -240,7 +242,7 @@ body {
   position: relative;
   overflow: hidden;
   border: 3px solid #fff;
-  z-index: 1;
+  z-index: -2;
 }
 
 .go-down {
@@ -821,7 +823,7 @@ body {
     align-items: center;
     list-style-type: none;
     width: 100%;
-    height: 70px;
+    height: 60px;
   }
 
   .header-mobile-text {
@@ -829,7 +831,7 @@ body {
     font-family: "Inter", sans-serif;
     font-style: normal;
     font-weight: 600;
-    margin: 20px;
+    margin: 50px;
     color: var(--text-color-2);
   }
