+++ index.html
@@ -209,9 +209,8 @@
     </footer>
   </section>
 
-  <script src="./js/script.js" type="module"></script>
+  <script src="./js/script.js" ></script>
   <script src="./js/form-validation.js" type="module"></script>
-  <script src="./js/animate.js" async></script>
 </body>
 
 </html>
